@@ -1,50 +1,39 @@
 {
   "name": "projectbluefire",
-  "version": "0.1.0",
   "private": true,
+  "version": "0.0.0",
+  "type": "commonjs",
   "scripts": {
-    "serve": "vue-cli-service serve",
-    "build": "vue-cli-service build",
-    "lint": "vue-cli-service lint"
+    "dev": "vite",
+    "build": "vite build",
+    "preview": "vite preview",
+    "lint": "eslint --ext .js,.vue --ignore-path .gitignore --fix src",
+    "format": "prettier .  --write"
   },
   "dependencies": {
-    "axios": "0.24.0",
-    "buttercms": "1.2.3",
-    "dayjs": "1.9.1",
-    "express": "4.17.1",
-    "firebase": "9.4.0",
-    "http2": "3.3.7",
-    "leaflet": "1.7.1",
-    "leaflet.markercluster": "1.4.1",
-    "nodemon": "2.0.4",
-    "socket.io-client": "2.3.1",
-    "vue": "2.6.12",
-    "vue-browser-geolocation": "1.8.0",
-    "vue-google-oauth2": "1.5.5",
-    "vue-router": "3.4.6",
-    "vue-socket.io-extended": "4.0.5",
-    "vue-uuid": "2.0.2",
-    "vue2-leaflet": "2.5.2",
-    "vue2-leaflet-movingmarker": "1.1.0",
-    "vuetify": "2.3.0",
-    "vuetify-google-autocomplete": "2.0.3",
-    "vuex": "3.5.1",
-    "yahoo-fantasy": "^4.1.5"
+    "@mdi/font": "7.4.47",
+    "@tanstack/vue-query": "5.17.11",
+    "@vueuse/core": "11.0.1",
+    "axios": "1.6.8",
+    "d3": "7.8.5",
+    "firebase": "10.11.1",
+    "firebase-functions": "4.5.0",
+    "firebaseui": "6.1.0",
+    "pinia": "2.1.7",
+    "vue": "3.5.3",
+    "vue-query": "1.26.0",
+    "vue-router": "4.3.0",
+    "vuetify": "3.7.0",
+    "dayjs": "1.11.11"
   },
   "devDependencies": {
-    "@fortawesome/fontawesome-free": "5.15.1",
-    "@vue/cli-plugin-babel": "4.5.7",
-    "@vue/cli-plugin-eslint": "4.5.7",
-    "@vue/cli-service": "4.5.7",
-    "axios": "0.24.0",
-    "babel-eslint": "10.1.0",
-    "eslint": "5.16.0",
-    "eslint-plugin-vue": "5.2.3",
-    "sass": "1.27.0",
-    "sass-loader": "7.3.1",
-    "socket.io": "2.3.0",
-    "vue-cli-plugin-vuetify": "1.1.2",
-    "vue-template-compiler": "2.6.12",
-    "vuetify-loader": "1.6.0"
+    "@vitejs/plugin-vue": "5.0.4",
+    "eslint": "8.56.0",
+    "eslint-config-prettier": "9.1.0",
+    "eslint-plugin-vue": "9.24.1",
+    "prettier": "3.2.5",
+    "sass": "1.69.5",
+    "vite": "5.2.8",
+    "vite-svg-loader": "5.1.0"
   }
 }
