@@ -3,15 +3,58 @@ module.exports = {
   env: {
     node: true
   },
-  'extends': [
-    'plugin:vue/essential',
-    // 'eslint:recommended'
-  ],
+  extends: ['eslint:recommended', 'plugin:vue/vue3-recommended', 'prettier'],
   rules: {
-    // 'no-console': process.env.NODE_ENV === 'production' ? 'error' : 'off',
-    // 'no-debugger': process.env.NODE_ENV === 'production' ? 'error' : 'off'
-  },
-  parserOptions: {
-    parser: 'babel-eslint'
+    'no-console': process.env.NODE_ENV === 'production' ? 'warn' : 'off',
+    'no-debugger': process.env.NODE_ENV === 'production' ? 'warn' : 'off',
+    'vue/component-api-style': [
+      'error',
+      ['script-setup', 'composition'] // "script-setup", "composition", "composition-vue2", or "options"
+    ],
+    'vue/new-line-between-multi-line-property': [
+      'error',
+      {
+        minLineOfMultilineProperty: 2
+      }
+    ],
+    'vue/component-tags-order': [
+      'error',
+      {
+        order: ['template', 'script', 'style']
+      }
+    ],
+    'vue/block-order': [
+      'error',
+      {
+        order: [['template', 'script'], 'style']
+      }
+    ],
+    'vue/padding-line-between-blocks': ['error', 'always'],
+    // Additional rules
+    "vue/html-closing-bracket-newline": [
+      "error",
+      {
+        "singleline": "never",
+        "multiline": "always",
+        "selfClosingTag": {
+          "singleline": "never",
+          "multiline": "always"
+        }
+      }
+    ],
+    'vue/attributes-order': ['error', {
+      'order': [
+        'DEFINITION',
+        'LIST_RENDERING',
+        'CONDITIONALS',
+        'RENDER_MODIFIERS',
+        'GLOBAL',
+        'UNIQUE',
+        'SLOT',
+        'EVENTS',
+        'OTHER_ATTR',
+        'CONTENT'
+      ]
+    }]
   }
 }
