@@ -1,11 +1,11 @@
 import Vue from 'vue'
 import Router from 'vue-router'
 import Home from './views/Home.vue'
-import SWF from "./views/SWF";
-import BlogHome from "./views/Blog/BlogHome";
-import BlogPost from "./views/Blog/BlogPost";
-import Portfolio from './views/Portfolio'
-import Merc from './views/Merc'
+import SWF from "./views/SWF.vue";
+import BlogHome from "./views/Blog/BlogHome.vue";
+import BlogPost from "./views/Blog/BlogPost.vue";
+import Portfolio from './views/Portfolio.vue'
+import Merc from './views/Merc.vue'
 import nhl21 from '@/views/nhl21'
 import DWF from '@/views/DWF'
 import Fantasy from '@/views/Fantasy'
